requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.9.0
asgiref==3.9.0
astor==0.8.1
asttokens==3.0.0
beautifulsoup4==4.14.2
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
dataclasses-json==0.6.7
decorator==4.4.2
distro==1.9.0
Django==5.2.4
django-admin-easy==0.8.0
django-admin-soft-dashboard==1.0.22
django-debug-toolbar==4.4.6
django-jazzmin==3.0.1
djangorestframework==3.15.2
executing==2.2.0
filelock==3.19.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.24.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instagrapi==2.2.1
instaloader==4.15a1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
linkedin-api==2.3.1
lxml==5.4.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
moviepy==1.0.3
mypy_extensions==1.1.0
mysqlclient==2.2.7
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
proglog==0.1.12
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-facebook-api==0.20.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==1.3.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
str2bool==1.1
tenacity==8.5.0
tokenizers==0.22.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
